+++ mullvad-management-interface/src/types/conversions/access_method.rs
@@ -11,7 +11,7 @@ mod settings {
                 direct: Some(settings.direct().clone().into()),
                 mullvad_bridges: Some(settings.mullvad_bridges().clone().into()),
                 encrypted_dns_proxy: Some(settings.encrypted_dns_proxy().clone().into()),
-                // domain_fronting: Some(settings.domain_fronting().clone().into()),
+                domain_fronting: Some(settings.domain_fronting().clone().into()),
                 custom: settings
                     .iter_custom()
                     .cloned()
@@ -46,12 +46,12 @@ mod settings {
                 ))
                 .and_then(access_method::AccessMethodSetting::try_from)?;
 
-            // let domain_fronting = settings
-            //     .domain_fronting
-            //     .ok_or(FromProtobufTypeError::invalid_argument(
-            //         "Could not deserialize Domain fronting Access Method from protobuf",
-            //     ))
-            //     .and_then(access_method::AccessMethodSetting::try_from)?;
+            let domain_fronting = settings
+                .domain_fronting
+                .ok_or(FromProtobufTypeError::invalid_argument(
+                    "Could not deserialize Domain fronting Access Method from protobuf",
+                ))
+                .and_then(access_method::AccessMethodSetting::try_from)?;
 
             let custom = settings
                 .custom
@@ -63,7 +63,7 @@ mod settings {
                 direct,
                 mullvad_bridges,
                 encrypted_dns_proxy,
-                // domain_fronting,
+                domain_fronting,
                 custom,
             ))
         }
@@ -139,7 +139,7 @@ mod data {
                 proto::access_method::AccessMethod::EncryptedDnsProxy(proxy) => {
                     AccessMethod::from(proxy)
                 }
-                // proto::access_method::AccessMethod::DomainFronting(df) => AccessMethod::from(df),
+                proto::access_method::AccessMethod::DomainFronting(df) => AccessMethod::from(df),
                 proto::access_method::AccessMethod::Custom(custom) => {
                     CustomProxy::try_from(custom).map(AccessMethod::from)?
                 }
@@ -174,11 +174,11 @@ mod data {
         }
     }
 
-    // impl From<proto::access_method::DomainFronting> for AccessMethod {
-    //     fn from(_value: proto::access_method::DomainFronting) -> Self {
-    //         AccessMethod::from(BuiltInAccessMethod::DomainFronting)
-    //     }
-    // }
+    impl From<proto::access_method::DomainFronting> for AccessMethod {
+        fn from(_value: proto::access_method::DomainFronting) -> Self {
+            AccessMethod::from(BuiltInAccessMethod::DomainFronting)
+        }
+    }
 
     impl TryFrom<proto::Socks5Local> for AccessMethod {
         type Error = FromProtobufTypeError;
@@ -225,11 +225,12 @@ mod data {
                     proto::access_method::AccessMethod::EncryptedDnsProxy(
                         proto::access_method::EncryptedDnsProxy {},
                     )
-                } // mullvad_types::access_method::BuiltInAccessMethod::DomainFronting => {
-                  //     proto::access_method::AccessMethod::DomainFronting(
-                  //         proto::access_method::DomainFronting {},
-                  //     )
-                  // }
+                }
+                mullvad_types::access_method::BuiltInAccessMethod::DomainFronting => {
+                    proto::access_method::AccessMethod::DomainFronting(
+                        proto::access_method::DomainFronting {},
+                    )
+                }
             }
         }
     }