+++ android/CHANGELOG.md
@@ -22,6 +22,9 @@ Line wrap the file at 100 chars.                                              Th
 * **Security**: in case of vulnerabilities.
 
 ## [Unreleased]
+### Added
+- Add a new access method: Domain fronting. Domain fronting is a way to reach the API via
+  proxies. The access method is enabled by default.
 
 
 ## [android/2026.9-beta1] - 2026-08-20